requirements.txt
django
djangorestframework
pillow
factory-boy
django-filter